Fixed test_validate.py
