+++ tests/collection/test_optional_members.py
@@ -5,13 +5,14 @@ from __future__ import annotations
 import polars as pl
 
 import pydiverse.colspec as cs
+import pydiverse.colspec.collection
 
 
 class TestColSpec(cs.ColSpec):
     a = cs.Integer()
 
 
-class MyCollection(cs.Collection):
+class MyCollection(pydiverse.colspec.collection.Collection):
     first: TestColSpec
     second: TestColSpec | None
