+++ tests/schema/test_validate.py
@@ -12,25 +12,38 @@ from pydiverse.colspec.exc import (
     RuleValidationError,
     ValidationError,
 )
+from pydiverse.colspec.optional_dependency import C, ColExpr, pdt
 
 
-class MySchema(cs.ColSpec):
+class MyColSpec(cs.ColSpec):
     a = cs.Int64(primary_key=True)
     b = cs.String(nullable=False, max_length=5)
     c = cs.String()
 
 
-class MyComplexSchema(cs.ColSpec):
+class MyComplexColSpec(cs.ColSpec):
     a = cs.Int64()
     b = cs.Int64()
 
-    @cs.rule()
+    @cs.rule_polars()
+    @staticmethod
     def b_greater_a() -> pl.Expr:
         return pl.col("b") > pl.col("a")
 
-    @cs.rule(group_by=["a"])
+    @cs.rule_polars(group_by=["a"])
+    @staticmethod
     def b_unique_within_a() -> pl.Expr:
         return pl.col("b").n_unique() == 1
+
+    @cs.rule()
+    @staticmethod
+    def b_greater_a2() -> ColExpr:
+        return C.b > C.a
+
+    @staticmethod
+    @cs.rule(group_by=["a"])
+    def b_unique_within_a2() -> ColExpr:
+        return C.b.count() == 1  # TODO: n_unique() is not available in pdt
 
 
 # -------------------------------------- COLUMNS ------------------------------------- #
@@ -40,8 +53,8 @@ class MyComplexSchema(cs.ColSpec):
 def test_missing_columns(df_type: type[pl.DataFrame] | type[pl.LazyFrame]):
     df = df_type({"a": [1], "b": [""]})
     with pytest.raises(ValidationError):
-        MySchema.validate_polars(df)
-    assert not MySchema.is_valid_polars(df)
+        MyColSpec.validate_polars(df)
+    assert not MyColSpec.is_valid_polars(df)
 
 
 # -------------------------------------- DTYPES -------------------------------------- #
@@ -51,20 +64,20 @@ def test_missing_columns(df_type: type[pl.DataFrame] | type[pl.LazyFrame]):
 def test_invalid_dtype(df_type: type[pl.DataFrame] | type[pl.LazyFrame]):
     df = df_type({"a": [1], "b": [1], "c": [1]})
     try:
-        MySchema.validate_polars(df)
+        MyColSpec.validate_polars(df)
         raise AssertionError()  # above should raise
     except DtypeValidationError as exc:
         assert len(exc.errors) == 2
-    assert not MySchema.is_valid_polars(df)
+    assert not MyColSpec.is_valid_polars(df)
 
 
 @pytest.mark.parametrize("df_type", [pl.DataFrame, pl.LazyFrame])
 def test_invalid_dtype_cast(df_type: type[pl.DataFrame] | type[pl.LazyFrame]):
     df = df_type({"a": [1], "b": [1], "c": [1]})
-    actual = MySchema.validate_polars(df, cast=True)
+    actual = MyColSpec.validate_polars(df, cast=True)
     expected = pl.DataFrame({"a": [1], "b": ["1"], "c": ["1"]})
     assert_frame_equal(actual, expected)
-    assert MySchema.is_valid_polars(df, cast=True)
+    assert MyColSpec.is_valid_polars(df, cast=True)
 
 
 # --------------------------------------- RULES -------------------------------------- #
@@ -74,36 +87,47 @@ def test_invalid_dtype_cast(df_type: type[pl.DataFrame] | type[pl.LazyFrame]):
 def test_invalid_column_contents(df_type: type[pl.DataFrame] | type[pl.LazyFrame]):
     df = df_type({"a": [1, 2, 3], "b": ["x", "longtext", None], "c": ["1", None, "3"]})
     try:
-        MySchema.validate_polars(df)
+        MyColSpec.validate_polars(df)
         raise AssertionError()  # above should raise
     except RuleValidationError as exc:
         assert len(exc.schema_errors) == 0
         assert exc.column_errors == {"b": {"nullability": 1, "max_length": 1}}
-    assert not MySchema.is_valid_polars(df)
+    assert not MyColSpec.is_valid_polars(df)
 
 
 @pytest.mark.parametrize("df_type", [pl.DataFrame, pl.LazyFrame])
 def test_invalid_primary_key(df_type: type[pl.DataFrame] | type[pl.LazyFrame]):
     df = df_type({"a": [1, 1], "b": ["x", "y"], "c": ["1", "2"]})
     try:
-        MySchema.validate_polars(df)
+        MyColSpec.validate_polars(df)
         raise AssertionError()  # above should raise
     except RuleValidationError as exc:
         assert exc.schema_errors == {"primary_key": 2}
         assert len(exc.column_errors) == 0
-    assert not MySchema.is_valid_polars(df)
+    assert not MyColSpec.is_valid_polars(df)
 
 
 @pytest.mark.parametrize("df_type", [pl.DataFrame, pl.LazyFrame])
-def test_violated_custom_rule(df_type: type[pl.DataFrame] | type[pl.LazyFrame]):
+def test_violated_custom_rule_polars(df_type: type[pl.DataFrame] | type[pl.LazyFrame]):
     df = df_type({"a": [1, 1, 2, 3, 3], "b": [2, 2, 2, 4, 5]})
     try:
-        MyComplexSchema.validate_polars(df)
+        MyComplexColSpec.validate_polars(df)
         raise AssertionError()  # above should raise
     except RuleValidationError as exc:
         assert exc.schema_errors == {"b_greater_a": 1, "b_unique_within_a": 2}
         assert len(exc.column_errors) == 0
-    assert not MyComplexSchema.is_valid_polars(df)
+    assert not MyComplexColSpec.is_valid_polars(df)
+
+
+def test_violated_custom_rule():
+    tbl = pdt.Table({"a": [1, 1, 2, 3, 3], "b": [2, 2, 2, 4, 5]})
+    try:
+        MyComplexColSpec.validate(tbl)
+        raise AssertionError()  # above should raise
+    except RuleValidationError as exc:
+        assert exc.schema_errors == {"b_greater_a2": 1, "b_unique_within_a2": 4}
+        assert len(exc.column_errors) == 0
+    assert not MyComplexColSpec.is_valid(tbl)
 
 
 @pytest.mark.parametrize("df_type", [pl.DataFrame, pl.LazyFrame])
@@ -113,9 +137,9 @@ def test_success_multi_row_strip_cast(
     df = df_type(
         {"a": [1, 2, 3], "b": ["x", "y", "z"], "c": [1, None, None], "d": [1, 2, 3]}
     )
-    actual = MySchema.validate_polars(df, cast=True)
+    actual = MyColSpec.validate_polars(df, cast=True)
     expected = pl.DataFrame(
         {"a": [1, 2, 3], "b": ["x", "y", "z"], "c": ["1", None, None]}
     )
     assert_frame_equal(actual, expected)
-    assert MySchema.is_valid_polars(df, cast=True)
+    assert MyColSpec.is_valid_polars(df, cast=True)
